Use C++11 for compilation Pretty compilation removed warning

+++ scintilla/gtk/ScintillaGTK.cxx
@@ -1258,7 +1258,7 @@ void ScintillaGTK::AddToPopUp(const char *label, int cmd, bool enabled) {
 	GtkItemFactoryEntry itemEntry = {
 	    fulllabel, NULL,
 	    menuSig,
-	    cmd,
+	    (guint)cmd,
 	    const_cast<gchar *>(label[0] ? "<Item>" : "<Separator>"),
 	    NULL
 	};
